+++ src/fdtdx/fdtd/initialization.py
